agent/requirements.txt
google-api-python-client>=2.0.0
google-adk>=0.1.0
google-generativeai>=0.8.0
python-dotenv>=1.1.0
requests>=2.32.4
rich>=14.0.0
pydantic>=2.11.7
langchain_core>=0.3.66
langchain-ollama>=0.3.3
langchain-huggingface>=0.3.0
langchain-chroma>=0.2.4
python-dotenv>=1.1.0
streamlit>=1.46.0
langchain_community>=0.3.26
langserve>=0.3.1
fastapi>=0.115.13
uvicorn>=0.34.3
sse_starlette>=2.3.6
bs4>=0.0.2
pypdf>=5.6.1
chromadb>=1.0.13
sentence-transformers>=4.1.0
langchain>=0.3.26
pillow>=11.2.1
supabase
flask
requests
flask_cors